+++ src/main/java/org/betonquest/betonquest/compatibility/quests/ConditionRequirement.java
@@ -1,21 +1,22 @@
 package org.betonquest.betonquest.compatibility.quests;
 
-import me.blackvein.quests.CustomRequirement;
+import me.pikamug.quests.module.BukkitCustomRequirement;
 import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.profiles.OnlineProfile;
 import org.betonquest.betonquest.exceptions.ObjectNotFoundException;
 import org.betonquest.betonquest.id.ConditionID;
 import org.betonquest.betonquest.utils.PlayerConverter;
-import org.bukkit.entity.Player;
+import org.bukkit.Bukkit;
 
 import java.util.Map;
+import java.util.UUID;
 
 /**
  * Requires the player to meet specified condition.
  */
 @SuppressWarnings("PMD.CommentRequired")
-public class ConditionRequirement extends CustomRequirement {
+public class ConditionRequirement extends BukkitCustomRequirement {
     /**
      * Custom {@link BetonQuestLogger} instance for this class.
      */
@@ -25,15 +26,15 @@ public class ConditionRequirement extends CustomRequirement {
         super();
         this.log = log;
         setName("BetonQuest condition");
-        setAuthor("Co0sh");
+        setAuthor("BetonQuest");
         addStringPrompt("Condition", "Specify BetonQuest condition name (with the package, like: package.condition)", null);
     }
 
     @Override
-    public boolean testRequirement(final Player player, final Map<String, Object> dataMap) {
+    public boolean testRequirement(final UUID uuid, final Map<String, Object> dataMap) {
         final String string = dataMap.get("Condition").toString();
         try {
-            final OnlineProfile onlineProfile = PlayerConverter.getID(player);
+            final OnlineProfile onlineProfile = PlayerConverter.getID(Bukkit.getPlayer(uuid));
             final ConditionID condition = new ConditionID(null, string);
             return BetonQuest.condition(onlineProfile, condition);
         } catch (final ObjectNotFoundException e) {